added tabs for addNewDataset and made component function for api calls
