+++ frontend/Components/AddNewDataset/ImportDataset.js
@@ -1,0 +1,22 @@
+import { Grid, Button, FormControl, FormLabel, Divider, Snackbar } from '@material-ui/core';
+
+
+export default function CreateDataset(){
+
+    return(
+        <Grid
+            container
+            spacing={1}
+            direction="column"
+            alignItems="center"
+            style={{ minHeight: '70vh', minWidth: '60vh', marginTop: "5vh"}}
+        >
+            <h1>import her</h1>
+
+
+        </Grid>
+    )
+
+}
+
+
