+++ frontend/Components/ApiCalls/GetApi.js
@@ -1,0 +1,16 @@
+
+
+const GetApi = async (url, setResponse) => {
+    try{
+        fetch(url, {
+            method: 'GET',    
+        })
+        .then(response => response.json())
+        .then(response => { setResponse(response); console.log(response)})
+    }
+    catch(_){
+        console.log("failed to fetch from: " + url)
+    }
+}
+
+export default GetApi;
